wip: tests schedule done

+++ graphql/app/jobs/remove_firm_job.rb
@@ -8,7 +8,7 @@ class RemoveFirmJob < ApplicationJob
     firm.events.each do |event|
       event.soft_delete!
       Schedule.where(event_id: event.id).left_outer_joins(:bookings).each do |schedule|
-        schedule.destroy
+        schedule.destroy if schedule.scheduled_for.to_datetime > Time.zone.now
       end
     end
   end
